requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Pillow>=10.3.0
python-dotenv==1.0.0
xlsxwriter==3.1.9
geopandas>=0.14.1
pyproj>=3.6.0
psycopg2-binary>=2.9.9
deep-translator>=1.11.4
gunicorn==21.2.0
email-validator==2.1.0
pandas>=2.0.0
openpyxl>=3.0.0
reportlab>=4.0.0
weasyprint>=60.0
dropbox>=12.0.2
